pessoa n desentegra mais

+++ Predio.h
@@ -16,7 +16,7 @@ typedef struct Andar {
 void construir_predio(Andar **predio, int andares);
 Andar* buscar_andar(Andar *predio, int valor);
 void inserir_f(Andar **predio, int valor);
-void chamar_elevador(struct Pessoa **fila, int tempo);
+void chamar_elevador(struct Pessoa **fila, unsigned long int tempo);
 void remover_andar(Andar **predio, int valor);
 void listar_andares(Andar *predio);
 void limpar_predio(Andar **predio);
